+++ doc/components.md
@@ -211,38 +211,6 @@ Check the example app `nine_tile_box` details on how to use it.
 
 ## Effects
 
-Flame provides a set of effects that can be applied to a certain type of components, these effects can be used to animate some properties of your components, check below for a list of those effects.
+Flame provides a set of effects that can be applied to a certain type of components, these effects can be used to animate some properties of your components, like position or dimensions. You can check the list of those effects [here](/doc/effects)
 
-Examples of the running effects can be found [here](/doc/examples/effects);
-
-### MoveEffect
-
-Applied to `PositionComponent`s, this effect can be used to move the component to a new position, using an [animation curve](https://api.flutter.dev/flutter/animation/Curves-class.html).
-
-Usage example:
-```dart
-import 'package:flame/effects/effects.dart';
-
-// Square is a PositionComponent
-square.addEffect(MoveEffect(
-  destination: Position(200, 200),
-  speed: 250.0,
-  curve: Curves.bounceInOut,
-));
-```
-
-### ScaleEffect
-
-Applied to `PositionComponent`s, this effect can be used to change the width and height of the component, using an [animation curve](https://api.flutter.dev/flutter/animation/Curves-class.html).
-
-Usage example:
-```dart
-import 'package:flame/effects/effects.dart';
-
-// Square is a PositionComponent
-square.addEffect(ScaleEffect(
-  size: Size(300, 300),
-  speed: 250.0,
-  curve: Curves.bounceInOut,
-));
-```
+Examples of the running effects can be found [here](/doc/examples/effects.md);
